Added Program Waitlists
Nearly identical to that of "serviceWaitlist", however it stores a programOccurrence instead of a serviceOccurrence, and a list of "programWaitlistEntry" instead of a "waitlistEntry". There are two different classes of models for program and service waitlists because it would not make sense to give both a program AND service occurrence to the same model and have one of these values be blank/null depending on which of the two is using it.